Rename 'frames' to 'chunks'.
The underlying Java library uses the term 'frame' to mean 'a set of
bytes that makes up the samples for all channels corresponding to one
point in time'. Which made the part of the code that deals with both
completely insane to read. Plus, this is more consistent with how it's
written in the sampled-sound implementation.